[AND-809] Remove unused code from apkfy companion multi install previous experiment
